some changes

+++ src/engine/webGPU/shds/merge_compute/merge_compute.wgsl
@@ -108,7 +108,7 @@ fn merge(cascadeIndex: f32, textCoords: vec2f) {
   for (var j: u32 = 0; j < 2; j++) {
     for (var k: u32 = 0; k < 2; k++) {
       var data: vec4f = textureLoad(resultTexture,
-                                    vec2u(u32(indexX1), u32(indexY1)) * u32(probe1Size) + vec2u(2 * pos.x + k, 2 * pos.y + j),
+                                    vec2u(u32(indexX1), u32(indexY1)) * u32(probe1Size) + vec2u(2 * u32(pos.x) + k, 2 * u32(pos.y) + j),
                                     u32(cascadeMaxIndex - cascadeIndex + 1));
       if (data.w != 0) {
         colors[0] += data.xyz;
@@ -125,7 +125,7 @@ fn merge(cascadeIndex: f32, textCoords: vec2f) {
   for (var j: u32 = 0; j < 2; j++) {
     for (var k: u32 = 0; k < 2; k++) {
       var data: vec4f = textureLoad(resultTexture,
-                                    vec2u(u32(indexX2), u32(indexY1)) * u32(probe1Size) + vec2u(2 * pos.x + k, 2 * pos.y + j),
+                                    vec2u(u32(indexX2), u32(indexY1)) * u32(probe1Size) + vec2u(2 * u32(pos.x) + k, 2 * u32(pos.y) + j),
                                     u32(cascadeMaxIndex - cascadeIndex + 1));
       if (data.w != 0) {
         colors[1] += data.xyz;
@@ -142,7 +142,7 @@ fn merge(cascadeIndex: f32, textCoords: vec2f) {
   for (var j: u32 = 0; j < 2; j++) {
     for (var k: u32 = 0; k < 2; k++) {
       var data: vec4f = textureLoad(resultTexture,
-                                    vec2u(u32(indexX1), u32(indexY2)) * u32(probe1Size) + vec2u(2 * pos.x + k, 2 * pos.y + j),
+                                    vec2u(u32(indexX1), u32(indexY2)) * u32(probe1Size) + vec2u(2 * u32(pos.x) + k, 2 * u32(pos.y) + j),
                                     u32(cascadeMaxIndex - cascadeIndex + 1));
       if (data.w != 0) {
         colors[2] += data.xyz;
@@ -159,7 +159,7 @@ fn merge(cascadeIndex: f32, textCoords: vec2f) {
   for (var j: u32 = 0; j < 2; j++) {
     for (var k: u32 = 0; k < 2; k++) {
       var data: vec4f = textureLoad(resultTexture,
-                                    vec2u(u32(indexX2), u32(indexY2)) * u32(probe1Size) + vec2u(2 * pos.x + k, 2 * pos.y + j),
+                                    vec2u(u32(indexX2), u32(indexY2)) * u32(probe1Size) + vec2u(2 * u32(pos.x) + k, 2 * u32(pos.y) + j),
                                     u32(cascadeMaxIndex - cascadeIndex + 1));
       if (data.w != 0) {
         colors[3] += data.xyz;
